[move] request / get/200_16をcgi/get/200_01として移動
