+++ test/common/request/cgi/get/200_01_print_ok_pl_body_message.txt
@@ -1,0 +1,7 @@
+GET /cgi-bin/print_ok.pl HTTP/1.1
+Host: localhost
+Connection: close
+Content-Length: 5
+Content-Type: text/plain
+
+HELLO
