--- test/common/request/get/2xx/200_16_body_message_cgi.txt
@@ -1,7 +1,0 @@
-GET /cgi-bin/print_ok.pl HTTP/1.1
-Host: localhost
-Connection: close
-Content-Length: 5
-Content-Type: text/plain
-
-HELLO
