+++ test/webserv/unit/http/test_case/test_case.hpp
@@ -29,7 +29,6 @@ int TestGetOk12HeaderFieldValueSpace(const server::VirtualServerAddrList &server
 int TestGetOk13SpaceHeaderFieldValue(const server::VirtualServerAddrList &server_infos);
 int TestGetOk14ExtraRequest(const server::VirtualServerAddrList &server_infos);
 int TestGetOk15BodyMessageDefault(const server::VirtualServerAddrList &server_infos);
-int TestGetOk16BodyMessageCgi(const server::VirtualServerAddrList &server_infos);
 int TestGetOk17NotExistHeaderField(const server::VirtualServerAddrList &server_infos);
 int TestGetOk18CgiScriptInPerl(const server::VirtualServerAddrList &server_infos);
 int TestGetOk19CgiScriptInPython(const server::VirtualServerAddrList &server_infos);
@@ -62,6 +61,9 @@ int TestGetMethodNotAllowed(const server::VirtualServerAddrList &server_infos);
 int TestGetTimeout1NoCrlf(const server::VirtualServerAddrList &server_infos);
 // 5xx
 int TestGetNotImplemented1NotExistMethod(const server::VirtualServerAddrList &server_infos);
+
+// cgi
+int TestCgiGetOk1PrintOkBodyMessage(const server::VirtualServerAddrList &server_infos);
 
 // GetErrorResponse
 int TestRequestTimeoutResponse();
